requirements.txt
groq>=0.9.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
